backend/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
firebase-admin
python-dotenv
